#8 Cleaning up the probe to use the new AppInstance and restoreable dock
