+++ applications/probe/src/main/java/org/phoebus/applications/probe/Probe.java
@@ -1,8 +1,14 @@
 package org.phoebus.applications.probe;
 
 import org.phoebus.framework.spi.AppDescriptor;
-import org.phoebus.framework.spi.AppInstance;
+import org.phoebus.ui.docking.DockItem;
+import org.phoebus.ui.docking.DockPane;
 
+/**
+ * 
+ * @author Kunal Shroff
+ *
+ */
 public class Probe implements AppDescriptor {
 
     public static final String NAME = "Probe";
@@ -18,9 +24,10 @@ public class Probe implements AppDescriptor {
     }
 
     @Override
-    public AppInstance create() {
-        //TODO
-        return null;
+    public ProbeInstance create() {
+        ProbeInstance probeInstance = new ProbeInstance(this);
+        DockPane.getActiveDockPane().addTab(new DockItem(probeInstance, probeInstance.create()));
+        return probeInstance;
     }
 
 }
